1. Fixed a clean-up script on mac. 2. Do not make a bak file when editing a bak file already

+++ qt_ui/mainwindow.cpp
@@ -80,6 +80,13 @@ namespace
             return true;
 
         QString bakFileName = fileInfo.dir().path() + "/" + fileInfo.completeBaseName() + QString(".bak");
+
+        if(bakFileName.toUpper() == fileName.toUpper())
+        {
+            // we are editing the bak file
+            return true;
+        }
+
         if(QFile::exists(bakFileName))
         {
             if(!QFile::remove(bakFileName))
